feat: refactor ProjectsSection to use custom hook for GitHub repos and add timeAgo utility
