+++ src/components/projects/LoadingSpinner.tsx
@@ -7,6 +7,6 @@ export const LoadingSpinner = () => (
     animate={{ opacity: 1 }}
     exit={{ opacity: 0 }}
   >
-    <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-red-500"></div>
+    <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-blue-400"></div>
   </motion.div>
 );
